requirements.txt
numpy
pandas
tensorflow
keras
seaborn
matplotlib
flask
uvicorn
pillow